# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SandBOX API Documentnation",
        "description": "Swagger OpenApi Documentnation",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products list",
                "description": "Get products list",
                "operationId": "productsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create new product",
                "description": "Create new product",
                "operationId": "productCreate",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "categories[]"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number"
                                    },
                                    "categories[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number"
                                    },
                                    "categories[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Title",
                                    "price": "100",
                                    "categories": [
                                        1,
                                        2
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The title field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product",
                "description": "Get product",
                "operationId": "productShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ProductId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product",
                "operationId": "productUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ProductId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "categories[]"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number"
                                    },
                                    "categories[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number"
                                    },
                                    "categories[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Title",
                                    "price": "100",
                                    "categories": [
                                        1,
                                        2
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The title field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "description": "Delete product",
                "operationId": "productDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ProductId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "description": "User identifier",
                                                    "type": "stringr",
                                                    "example": "deleted"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CategoryResource": {
                "title": "Category",
                "description": "API category item",
                "properties": {
                    "id": {
                        "description": "Identifier",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "description": "Categoty title",
                        "type": "string",
                        "example": "Title"
                    },
                    "created_at": {
                        "description": "Create date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-25T10:47:23.000000Z"
                    },
                    "updated_at": {
                        "description": "Update date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-25T10:47:23.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoryResource"
                }
            },
            "ProductResource": {
                "title": "Product",
                "description": "API product item",
                "properties": {
                    "id": {
                        "description": "Identifier",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "description": "Product title",
                        "type": "string",
                        "example": "Title"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "example": "100.55"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        }
                    },
                    "created_at": {
                        "description": "Create date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-25T10:47:23.000000Z"
                    },
                    "updated_at": {
                        "description": "Update date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-25T10:47:23.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductResource"
                }
            }
        }
    }
}